Stop weird shifting of rectangles in for loops
Using Math.round is not good because sometimes the rectangle shifts to the right side, other times to the left. The behavior of Processingjs is thus inconsistent with Processing in this respect, such that float coordinates iterated through a for loop will create overlapping rectangles or rectangles that are separated by a pixel.